Fix getNearestEpharmacyWarehouseLocation schema && set default visible layout to false

+++ features/merchant/shop_page/src/main/java/com/tokopedia/shop/info/domain/GetNearestEpharmacyWarehouseLocationUseCase.kt
@@ -27,10 +27,10 @@ class GetNearestEpharmacyWarehouseLocationUseCase @Inject constructor(
 
         fun createParams(
             shopId: Long,
-            districtId: Int
+            districtId: Long
         ): RequestParams = RequestParams.create().apply {
             putLong(SHOP_ID, shopId)
-            putInt(DISTRICT_ID, districtId)
+            putLong(DISTRICT_ID, districtId)
         }
     }
